@@ -5,7 +5,7 @@ import com.typesafe.config.Config;
 import dagger.Module;
 import dagger.Provides;
 
-import javax.inject.Singleton;
+import jakarta.inject.Singleton;
 
 @Module
 public class DaggerConfModule {
