@@ -1,8 +1,8 @@
 package com.coreoz.plume.scheduler;
 
-import javax.inject.Inject;
-import javax.inject.Provider;
-import javax.inject.Singleton;
+import jakarta.inject.Inject;
+import jakarta.inject.Provider;
+import jakarta.inject.Singleton;
 
 import com.coreoz.plume.services.time.TimeProvider;
 import com.coreoz.wisp.Scheduler;
