@@ -5,15 +5,14 @@ import java.io.InputStream;
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Type;
 
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.MultivaluedMap;
+import jakarta.ws.rs.core.MediaType;
+import jakarta.ws.rs.core.MultivaluedMap;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import lombok.extern.slf4j.Slf4j;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.jaxrs.cfg.Annotations;
-import com.fasterxml.jackson.jaxrs.json.JacksonJaxbJsonProvider;
+import com.fasterxml.jackson.jakarta.rs.cfg.Annotations;
+import com.fasterxml.jackson.jakarta.rs.json.JacksonJsonProvider;
 
 /**
  * A Jackson JSON provider that throws a {@link JsonRequestParseException}
@@ -21,9 +20,8 @@ import com.fasterxml.jackson.jaxrs.json.JacksonJaxbJsonProvider;
  * It is useful to return proper 400 errors when JSON request input is not valid
  * @see JacksonJaxbJsonProvider
  */
-public class WsJacksonJsonProvider extends JacksonJaxbJsonProvider {
-
-	private static final Logger logger = LoggerFactory.getLogger(WsJacksonJsonProvider.class);
+@Slf4j
+public class WsJacksonJsonProvider extends JacksonJsonProvider {
 
 	public WsJacksonJsonProvider() {
 		super();
@@ -48,5 +46,4 @@ public class WsJacksonJsonProvider extends JacksonJaxbJsonProvider {
 			throw new JsonRequestParseException();
 		}
 	}
-
 }
